Move code from AbstractTask to DefaultTask
and fix task creation
and (de)serialization by the configuration cache
