+++ subprojects/core/src/main/java/org/gradle/api/internal/project/taskfactory/TaskFactory.java
@@ -19,7 +19,6 @@ import org.gradle.api.DefaultTask;
 import org.gradle.api.Describable;
 import org.gradle.api.InvalidUserDataException;
 import org.gradle.api.Task;
-import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.internal.project.ProjectInternal;
 import org.gradle.api.reflect.ObjectInstantiationException;
 import org.gradle.api.tasks.TaskInstantiationException;
@@ -51,6 +50,9 @@ public class TaskFactory implements ITaskFactory {
     @Override
     @SuppressWarnings("deprecation")
     public <S extends Task> S create(final TaskIdentity<S> identity, @Nullable final Object[] constructorArgs) {
+
+        NameValidator.validate(identity.name, "task name", "");
+
         if (!Task.class.isAssignableFrom(identity.type)) {
             throw new InvalidUserDataException(String.format(
                 "Cannot create task '%s' of type '%s' as it does not implement the Task interface.",
@@ -58,28 +60,21 @@ public class TaskFactory implements ITaskFactory {
                 identity.type.getSimpleName()));
         }
 
-        NameValidator.validate(identity.name, "task name", "");
-
         final Class<? extends DefaultTask> implType;
         if (identity.type == Task.class) {
             implType = DefaultTask.class;
         } else if (DefaultTask.class.isAssignableFrom(identity.type)) {
             implType = identity.type.asSubclass(DefaultTask.class);
-        } else if (identity.type == org.gradle.api.internal.AbstractTask.class || identity.type == TaskInternal.class) {
-            throw new InvalidUserDataException(String.format(
-                "Cannot create task '%s' of type '%s' as this type is not supported for task registration.",
-                identity.identityPath.toString(),
-                identity.type.getSimpleName()));
         } else {
             throw new InvalidUserDataException(String.format(
-                "Cannot create task '%s' of type '%s' as directly extending AbstractTask is not supported.",
+                "Cannot create task '%s' of type '%s' as it does not extend DefaultTask.",
                 identity.identityPath.toString(),
                 identity.type.getSimpleName()));
         }
 
         Describable displayName = Describables.withTypeAndName("task", identity.getIdentityPath());
 
-        return org.gradle.api.internal.AbstractTask.injectIntoNewInstance(project, identity, new Callable<S>() {
+        return DefaultTask.injectIntoNewInstance(project, identity, new Callable<S>() {
             @Override
             public S call() {
                 try {
@@ -87,7 +82,7 @@ public class TaskFactory implements ITaskFactory {
                     if (constructorArgs != null) {
                         instance = instantiationScheme.instantiator().newInstanceWithDisplayName(implType, displayName, constructorArgs);
                     } else {
-                        instance = instantiationScheme.deserializationInstantiator().newInstance(implType, org.gradle.api.internal.AbstractTask.class);
+                        instance = instantiationScheme.deserializationInstantiator().newInstance(implType, DefaultTask.class);
                     }
                     return identity.type.cast(instance);
                 } catch (ObjectInstantiationException e) {
