+++ subprojects/core/src/test/groovy/org/gradle/api/internal/project/taskfactory/TaskFactoryTest.groovy
@@ -19,7 +19,6 @@ package org.gradle.api.internal.project.taskfactory
 import org.gradle.api.DefaultTask
 import org.gradle.api.InvalidUserDataException
 import org.gradle.api.Task
-import org.gradle.api.internal.AbstractTask
 import org.gradle.api.internal.TaskInternal
 import org.gradle.api.reflect.ObjectInstantiationException
 import org.gradle.api.tasks.TaskInstantiationException
@@ -76,7 +75,7 @@ class TaskFactoryTest extends AbstractProjectBuilderSpec {
         Task task = taskFactory.create(new TaskIdentity(TestDefaultTask, 'task', null, Path.path(':task'), null, 12), (Object[]) null)
 
         then:
-        1 * deserializeInstantiator.newInstance(TestDefaultTask, AbstractTask) >> { new TestDefaultTask() }
+        1 * deserializeInstantiator.newInstance(TestDefaultTask, DefaultTask) >> { new TestDefaultTask() }
         task instanceof TestDefaultTask
     }
 
@@ -95,19 +94,10 @@ class TaskFactoryTest extends AbstractProjectBuilderSpec {
 
         then:
         InvalidUserDataException e = thrown()
-        e.message == "Cannot create task ':task' of type '$taskType.simpleName' as this type is not supported for task registration."
+        e.message == "Cannot create task ':task' of type '$taskType.simpleName' as it does not extend DefaultTask."
 
         where:
-        taskType << [AbstractTask, TaskInternal]
-    }
-
-    void testCreateTaskForTypeDirectlyExtendingAbstractTask() {
-        when:
-        taskFactory.create(new TaskIdentity(ExtendsAbstractTask, 'task', null, Path.path(':task'), null, 12))
-
-        then:
-        InvalidUserDataException e = thrown()
-        e.message == "Cannot create task ':task' of type 'ExtendsAbstractTask' as directly extending AbstractTask is not supported."
+        taskType << [TaskInternal, NotADefaultTask]
     }
 
     void wrapsFailureToCreateTaskInstance() {
@@ -134,6 +124,6 @@ class TaskFactoryTest extends AbstractProjectBuilderSpec {
     static class NotATask {
     }
 
-    static class ExtendsAbstractTask extends AbstractTask {
+    static abstract class NotADefaultTask implements Task {
     }
 }
